+++ backend/src/main/java/ca/metricalsky/winston/service/ChannelService.java
@@ -1,6 +1,7 @@
 package ca.metricalsky.winston.service;
 
 import ca.metricalsky.winston.dto.ChannelDto;
+import ca.metricalsky.winston.exception.AppException;
 import ca.metricalsky.winston.mapper.dto.ChannelDtoMapper;
 import ca.metricalsky.winston.repository.ChannelRepository;
 import lombok.RequiredArgsConstructor;
@@ -8,7 +9,6 @@ import org.mapstruct.factory.Mappers;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Comparator;
 import java.util.List;
@@ -40,7 +40,7 @@ public class ChannelService {
 
     public void requireChannelExists(String channelId) {
         if (!channelRepository.existsById(channelId)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "The requested channel was not found.");
+            throw new AppException(HttpStatus.NOT_FOUND, "The requested channel was not found.");
         };
     }
 }