+++ frontend/src/routes/videos/id/index.tsx
@@ -10,6 +10,7 @@ import {FetchCommentsAlert} from "./FetchCommentsAlert";
 import {useAppSelector} from "../../../store/hooks";
 import {DateTime} from "luxon";
 import {descBy} from "../../../utils";
+import {CommentsDisabledJumbotron} from "./CommentsDisabledJumbotron";
 
 export const VideosIdRoute = () => {
     const {videoId} = useParams();
@@ -22,11 +23,11 @@ export const VideosIdRoute = () => {
     const {data: video} = useFindVideoByIdQuery(videoId!)
 
     const {data: comments} = useListCommentsByVideoIdQuery(videoId!)
-    const fetchedComments = useAppSelector(state => state.fetches.comments[videoId!]?.data)
+    const fetchState = useAppSelector(state => state.fetches.comments[videoId!])
     const combinedComments = useMemo(
-        () => (fetchedComments ?? []).concat(comments ?? [])
+        () => (fetchState?.data ?? []).concat(comments ?? [])
             .sort(descBy(comment => DateTime.fromISO(comment.publishedAt).valueOf())),
-        [comments, fetchedComments]
+        [comments, fetchState]
     );
 
     const displayedComments = useMemo(
@@ -38,6 +39,11 @@ export const VideosIdRoute = () => {
                     reply.text.toLowerCase().includes(search.toLowerCase())).length > 0)
             .slice(pageSize * (page - 1), pageSize * page) ?? [],
         [combinedComments, pageSize, page, search]
+    );
+
+    const commentsDisabled = useMemo(
+        () => video?.commentsDisabled || fetchState?.error?.type === "/api/problem/comments-disabled",
+        [video, fetchState]
     );
 
     return (
@@ -56,8 +62,9 @@ export const VideosIdRoute = () => {
                 </>}
             </Breadcrumb>
             {video && <VideoDetails video={video}/>}
-            {video && (combinedComments?.length ?? 0) == 0 && <NoCommentsJumbotron video={video}/>}
+            {video && !commentsDisabled && (combinedComments?.length ?? 0) == 0 && <NoCommentsJumbotron video={video}/>}
             {video && <FetchCommentsAlert video={video}/>}
+            {commentsDisabled && <CommentsDisabledJumbotron/>}
             {(combinedComments?.length ?? 0) > 0 && (
                 <>
                     <PaginationRow